chore(controller/certificates): move cert helpers to certificates.go
